PR feedback: colocate right-pane types and split shared layout files.
Move types next to their owners, rename the mode helper to withRightPane,
split the shared host into hook + root, and drop extra waitFors after close.
