+++ src/components/Layout/GridTableLayout/GridTableLayout.tsx
@@ -27,8 +27,7 @@ import { useTestIds } from "src/utils/useTestIds";
 import { zIndices } from "src/utils/zIndices";
 import { type BaseQueryTableProps, type GridTablePropsWithRows, isGridTableProps } from "../layoutTypes";
 import { DocumentScrollOverlayRightPaneLayout } from "../RightPaneLayout/DocumentScrollOverlayRightPaneLayout";
-import { resolveWithRightPaneOptions } from "../RightPaneLayout/documentScrollRightPaneMode";
-import type { WithRightPane } from "../RightPaneLayout/types";
+import { resolveWithRightPaneOptions, type WithRightPane } from "../RightPaneLayout/withRightPane";
 import { GridTableLayoutActions, type SearchBoxApi } from "./GridTableLayoutActions";
 import { QueryTable, type QueryTableProps } from "./QueryTable";
 import { usePersistedTableView } from "./usePersistedTableView";
