--- src/components/Layout/RightPaneLayout/types.ts
@@ -1,30 +1,0 @@
-import type { ReactNode } from "react";
-
-/** Desktop right-pane strategy. See `docs/layouts.md` for overlay / push / auto outcomes. */
-export type DocumentScrollRightPaneMode = "auto" | "overlay" | "push";
-
-/** Desktop strategies for {@link DocumentScrollInlineRightPaneLayout} only. */
-export type DocumentScrollInlineRightPaneMode = "auto" | "push";
-
-/** Default document-scroll detail pane width (px). */
-export const defaultDocumentScrollRightPaneWidth = 450;
-
-/** DOM marker on open pane content; polled by {@link waitForRightPaneExit} after close. */
-export const rightPaneContentDataAttribute = "data-right-pane-content";
-
-/**
- * Opt into `DocumentScrollOverlayRightPaneLayout`.
- * `true` / a px width use the caller's default mode; an object sets width and/or mode.
- */
-export type WithRightPane =
-  | boolean
-  | number
-  | {
-      width?: number;
-      mode?: DocumentScrollRightPaneMode;
-    };
-
-/** Body passed to `openRightPane`. */
-export type OpenRightPaneOpts = {
-  content: ReactNode;
-};
