+++ src/index.ts
@@ -55,8 +55,8 @@ export * from "./components/Layout/PreventBrowserScroll";
 export * from "./components/Layout/RightPaneLayout/DocumentScrollOverlayRightPaneLayout";
 export * from "./components/Layout/RightPaneLayout/RightPaneLayout";
 export * from "./components/Layout/RightPaneLayout/rightPaneStore";
-export * from "./components/Layout/RightPaneLayout/types";
 export * from "./components/Layout/RightPaneLayout/useRightPane";
+export * from "./components/Layout/RightPaneLayout/withRightPane";
 export * from "./components/Layout/ScrollableContent";
 export * from "./components/Layout/ScrollableFooter";
 export * from "./components/Layout/ScrollableParent";
